Adding perfstubs support to build (#1)
* Adding perfstubs as a submodule and build dependency

* Adding perfstubs directory for submodule?

* Adding comments to CMakeLists.txt file

* updating perstubs state

* updated perfstubs

* Debugging after testing on Summit
With these changes, XGC can be built with this camtimers with
perfstubs support without any modification in the XGC configuration.

* Adding perfstubs as a submodule and build dependency

Adding perfstubs directory for submodule?

Adding comments to CMakeLists.txt file

updating perstubs state

updated perfstubs

Debugging after testing on Summit
With these changes, XGC can be built with this camtimers with
perfstubs support without any modification in the XGC configuration.

* Restoring cast statement

* Adding required OpenMP flags and LINUX definition

* Removed submodule

* Changing perfstubs to be a subdirectory
If the subdirectory doesn't exist, git will clone it.  It is not a
submodule.  The config.h header for the perfstubs library is generated.
Then the lone perfstubs_api/timer.c file is builat as part of the camtimers
library.